refactor: use chrono::TimeDelta and move find_tokens_expiring_within_days out
